Add in basic tooling for CMake builds (#1)
This commit adds in some basic infrastructure for doing builds and
specifically creates a CMake builder. A corpus description for LLVM is
added. Note that this doesn't include any tooling to extract a bitcode
corpus into the ml-compiler-opt format. That will come later (probably
after some refactoring of the code there so that we can call it
programatically).