Add tooling for extracting IR (#8)
In order to build an IR corpus, we need to be able to extract IR from a
project after building it. This patch adds in support for extracting IR
using tooling from ml-compiler-opt. This patch also removes the
ml-compiler-opt submodule as this patch requires ml-compiler-opt to be
installed as a python package and support has recently landed upstream
to support that method of installation with PyPI packages coming soon.