Add support for not having a commit SHA
In a lot of cases it doesn't make sense to checkout a specific commit
SHA in a project that we're building. This commit makes it so that the
infrastructure used for downloading source code from git can work
without a commit SHA and does so more optimally by setting --depth=1.