Moved corpus directory construction and added logging to cargo builder.
This PR moves corpus directory construction so that logs and a build
manifest can be written even if there is a total build failure/nothing
has been extracted yet. This patch also adds some exported logs to the
cargo builder so that failures are nicely compartmentalized per target
and can be investigated after the fact by simply going through the log.
This also adds a preliminary build manifest (#11) that will be
significantly expanded in the future that includes log references and
whether or not the target was successfully built.