Tar archive source (#41)
* Make parse_crates_database.py also output tar archives

This patch makes parse_crates_database.py also give URLs to .tar.gz
archives for the latest version of each package. This also changes the
output format of the tool to JSON.

* Formatting fix

* Refactor git source to a separate module

* Add tar source

* Refactor source downloads to a separate module

This commit refactors source downloads to a seprate module and also
enables the use of the tar archive source downloader.

* Enable source fallbacks

This patch enables source failover where if one source doesn't work, the
next one in the list will be tried. This is primarily intended for
building rust crates where the repository is preferred but might not be
available.