Spack dependency builder (#52)
* Initial spack build command line tool

* added test thing

* Seemed to have gotten environment to concretize

* Many improvements to get_spack_package_list.py

* Get basic spack build infrastructure working

* Minor changes to get things working better

* Minor change to logging format

Significantly changes the way we're building spack packages (or at least a series of packages) and adds in the tooling to do so. Instead of using a build cache (which rarely hit and seemed to be pretty finnicky), now we just construct the transitive closure for a package. This still needs more work to bring it completely up to speed, but it allows us to build significantly more than before so that we can start to get some initial results.