Make get_spack_package_list.py go through all packages
Before it would only emit all the concretized version (along with
dependencies) for a single package. This changes that behavior to
actually write them all to a file and also changes the concretization
strategy so everything can actually be concretized with clang.

Still some work to do as Spack things some packages that can't be built
with clang can be built with clang, but we're definitely making some
progress.