Get spack staging directory by hash
We get the spack staging directory currently just using the name of the
program. This is not particularly scalable as mentioned in #55. This
patch changes it so that we access the staging directory by hash, which
we have in the manifest and is specific to the spec.

Fixes #55.
Also Fixes #56 (somehow, not sure what happened there or what this does
to fix it).