Bump ml-compiler-opt dep
This is so that we have access to the recent changes allowing
suppression of common warning/logging messages during the IR extraction
process.

This touches a couple other dependencies that should be fine but I
haven't done extensive validation yet. The only major one that I noticed
was the grpcio bump, but we only had problems with versions prior to
1.53.1, so the bump to the 1.56 series should be fine.