Change julia package server and adjust thread count
Adjusting the thread count primarily to combat high memory pressure
(some more testing is needed, but it seems to be almost stable on a 96
thread/256GB node with 1 thread per job). Also change the julia package
server to none so that stuff is pulled from sources as the julia package
server is heavily rate limited.