Add error handling for missing build directory and comment
I was running into some errors because of missing build directories
(more investigation should probably be done into why they're missing in
the first place), so this adds in some error handling for that. Also add
in a comment with a TODO to make this code work on multiple systems.