Add garbage collection timeout
There are a couple cases where I suspect Spack is getting into some
erroneous state and erroring out but not killing its own process, so
it hangs on garbage collection. This patch adds a timeout to hopefully
prevent that issue from impacting builds, and is good practice anyways
if that wasn't the main issue. The timeout is set pretty generously to
not interfere with any actual collection but short enough to time out
fairly quickly.