Llvm fork in container (#97)
* Switch container build to LLVM fork

This will allow us to take advantage of new features that we're
developing faster while we're in the process of upstreaming them.

* Remove LLVM commit sha from container build

We control this in the branch on our fork so we don't need to control
this explicitly within the container. This makes some things a little
bit easier (like setting --depth=1).