Make symlinks absolute for compiler wrapper
Without using absolute paths, it seems like the symlinks aren't
executable. Theoretically, this should work, but it should be fine to
keep them in absolute paths with the way the current infrastructure is
setup (laying everything out in the same folder structure everytime).
This will need to be fixed in the future though to make things more
resilient.