Fix error with length in c++ mode
This patch fixes an issue in c++ mode where if the compiler is invoked
without a -o flag (like to get the version), the length of the mode was
more than 1, which caused the wrapper to try and save source files and
fail as there aren't any specified. This patch fixes this by making the
mode still be a tuple.